training/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
scikit-learn>=1.3.0
onnx>=1.14.0
onnxruntime>=1.15.0
datasets>=2.12.0
accelerate>=0.20.0